requirements.txt
Django
mysqlclient
redis
pandas
openpyxl
requests
python-dotenv
loguru
django-redis
django-simpleui
django-unfold